Address comments from James
- do not return errors for invalid rowCountPtr and columnCountPtr
- add static cast for columnCount
- Add tests with SQL_C_DEFAULT as target type
- Add test checks for indicator
- Fix SQL_NUMERIC, SQL_BIT, SQL_FLOAT target type

+++ cpp/src/arrow/flight/sql/odbc/tests/odbc_test_suite.cc
@@ -124,7 +124,7 @@ std::wstring FlightSQLODBCRemoteTestBase::getQueryAllDataTypes() {
 
           CAST(-2147483648 AS INTEGER) AS sinteger_min,
           CAST(2147483647 AS INTEGER) AS sinteger_max,
-          CAST(0 AS INTEGER) AS uinteger_min,
+          CAST(0 AS BIGINT) AS uinteger_min,
           CAST(4294967295 AS BIGINT) AS uinteger_max,
 
           CAST(-9223372036854775808 AS BIGINT) AS sbigint_min,
